remove registry processing from filebeat check
This change removes the processing of the registry file from the
filebeat check. In v7.9.0, filebeat changed the format of the registry
file: https://github.com/elastic/beats/issues/21158. In a following
commit they communicate that they have not documented this change as
they view this as an internal API.

Rather than continue to process the registry for the
`filebeat.registry.unprocessed_bytes` metric, this change removes that
processing all together.